Do not enforce TLS on non-loopback RPC listen addresses
There are many valid configurations where TLS is not needed to secure
access to dcrd's RPC server, even when it is not listening on a
loopback interface.

Alternatives to securing and authenticating RPC without TLS:

- Listening on a virtual ethernet device, with strict firewall rules
  which enforce only authorized clients (like a user or group) are able
  to connect, and only from the same machine.

- Tunneling access over a (possibly public) physical network through a
  another encrypted and authenticated channel, e.g ssh or WireGuard.

In fact, some configurations can arguably provide or make it easier to
implement better security than the application-level security that
dcrd can provide itself through TLS.  This is possible when solutions
take advantage of operating system-level permissions, network
isolation, and packet filtering.  As an example, packet filtering is
quite often easier to describe and implement when filtering a virtual
network interface that is separate and isolated from any physical
egress or loopback interface.  Virtual network topologies can also
limit the scope of damage possible in case any process on the system
is compromised.

Of course, none of these benefits require TLS to be removed, and it is
still possible to build intricate and secure virtual network
topologies while keeping TLS enabled.  However, TLS may only be
providing diminishing returns at this point, and it is completely
reasonable to avoid its use in order to simplify the overal system
configuration if the proper precautions have been applied.

I currently have no interest in upstreaming this patch, and neither is
upstream interested in taking it.  Enforcing TLS in the default and
supported Decred releases has certainly forced the hand of many users
to improve their overall security situation, and for that I am
thankful.  This is not a patch for the masses or those who want to be
lax about security; rather it is for those who understand what they
are doing, and are willing to push the boundaries of deploying a
secure system using the best tools available: those tools which are
provided by their operating system and which cannot be sidestepped by
a compromised process.